pyproject.toml
[project]
name = "employee-management-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Ирина Демьяненко",email = "idemyanenko5@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (==4.2.*)",
    "djangorestframework (==3.*)",
    "django-filter (==24.*)",
    "drf-spectacular (==0.*)",
    "openpyxl (==3.*)",
    "python-dotenv (==1.*)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "django-stubs (>=5.2.7,<6.0.0)",
    "djangorestframework-stubs (>=3.16.4,<4.0.0)"
]
